+++ include/newton.h
@@ -18,12 +18,12 @@ int Newton(int (*vecfun)(double * y, double * Fy, void *param),
 	   double rtol,
 	   int *iter,
 	   int maxiter,
-	   int neq,
+	   size_t neq,
 	   ErrorMsg error_message);
 int jacobian_for_Newton(int (*vecfun)(double * y, double * Fy, void *param),
 			double *y0,
 			double *Fval,
-			int neq,
+			size_t neq,
 			void *param,
 			double **jac);
