+++ tools/evolver_rk45.c
@@ -5,7 +5,7 @@ int evolver_rk45(int (*derivs)(double x,double * y,double * dy,
 		 double x_ini,
 		 double x_final,
 		 double * y_inout, 
-		 int neq, 
+		 size_t neq, 
 		 EvolverOptions *options,
 		 ErrorMsg error_message){
 	
@@ -241,7 +241,7 @@ int evolver_rkdp45(int (*derivs)(double x,
 		   double t_ini,
 		   double t_final,
 		   double * y_inout, 
-		   int neq, 
+		   size_t neq, 
 		   EvolverOptions *options,
 		   ErrorMsg error_message){
